helpers for gc analysis
